+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 QuantumCircuitOpt.jl Change Log
 ===============================
 
+### v0.6.2
+- Re-implemented post-optimization circuit compression algorithm - generalizes from `2` qubits to `n` qubits, and to gates with angle parameters. 
+- Re-implemented `kron_layer` with a left-to-right wire scanner. It prints minimal tensor expressions (`CNot_{1,2}`, `X ⊗ X`) with `⊗ I` factors for any register size.
+- Cleaner logging of input gates and optimal circuit
+- Clean up of functions in `src/utility.jl`
+
 ### v0.6.1
 - Added support for T-gate count minimization
 - Cleaned up `src/utility.jl`